Fixed a bug where the ShootingAI could choose weapons that weren't purchased because of magic number usage.
Fixed a bug where the ShootingAI would sometimes shoot a weapon other than the Tractor Beam at a Shard.

ShootingAI gives higher priority to Shards when the ship is hurt.

Increased the weighting of the railgun in certain situations so we can see it used more.

Minor speed optimizations to the ShootingAI. Commented some code.


git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@944 342171fe-4ba3-422f-b357-09b6f24955fe